replaced interface with abstract class
