+++ src/Basket/Comment.php
@@ -3,15 +3,28 @@
  * @copyright   Copyright (c) Wizacha
  * @license     Proprietary
  */
-declare(strict_types=1);
+declare(strict_types = 1);
 
 
 namespace Wizaplace\SDK\Basket;
 
-interface Comment
+use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
+
+abstract class Comment
 {
-    /**
-     * Serialize Comment objects.
-     */
-    public function toArray(): array;
+    /** @var string */
+    protected $comment;
+
+    public function __construct(string $comment)
+    {
+        if ($comment === '') {
+            throw new SomeParametersAreInvalid('Missing comment');
+        }
+        $this->comment = $comment;
+    }
+
+    public function getComment(): string
+    {
+        return $this->comment;
+    }
 }
