+++ src/Basket/ProductComment.php
@@ -11,39 +11,26 @@ namespace Wizaplace\SDK\Basket;
 
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 
-class ProductComment implements Comment
+class ProductComment extends Comment
 {
     /**
      * @var string
      */
     private $declinationId;
 
-    /**
-     * @var string
-     */
-    private $comment;
-
     public function __construct(string $declinationId, string $comment)
     {
+        parent::__construct($comment);
         if ($declinationId === '') {
             throw new SomeParametersAreInvalid('Missing declination Id');
         }
-        if ($comment === '') {
-            throw new SomeParametersAreInvalid('Missing comment');
-        }
 
         $this->declinationId = $declinationId;
-        $this->comment = $comment;
     }
 
     public function getDeclinationId(): string
     {
         return $this->declinationId;
-    }
-
-    public function getComment(): string
-    {
-        return $this->comment;
     }
 
     public function toArray(): array
