+++ src/Basket/ShippingGroupComment.php
@@ -11,36 +11,26 @@ namespace Wizaplace\SDK\Basket;
 
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 
-class ShippingGroupComment implements Comment
+class ShippingGroupComment extends Comment
 {
     /**
      * @var int
      */
     private $shippingGroupId;
 
-    /**
-     * @var string
-     */
-    private $comment;
-
     public function __construct(int $shippingGroupId, string $comment)
     {
+        parent::__construct($comment);
         if ($comment === '') {
             throw new SomeParametersAreInvalid('Missing comment');
         }
 
         $this->shippingGroupId = $shippingGroupId;
-        $this->comment = $comment;
     }
 
     public function getShippingGroupId(): int
     {
         return $this->shippingGroupId;
-    }
-
-    public function getComment(): string
-    {
-        return $this->comment;
     }
 
     public function toArray(): array
